Added logging

+++ DatabaseUpdater/Engine.cs
@@ -6,6 +6,7 @@ using Microsoft.Win32;
 using System.Windows.Forms;
 using System.Diagnostics.Contracts;
 using System.Data.Odbc;
+using System.Reflection;
 
 namespace DatabaseUpdater
 {
@@ -13,31 +14,53 @@ namespace DatabaseUpdater
     {
         private MainForm _Form;
 
-        public Engine(MainForm TheForm) 
+        public Engine(MainForm TheForm)
         {
             _Form = TheForm;
         }
 
         public bool Run()
         {
-
+            if (_Form.IsLoggingOn)
+                _Form.log.WriteLine($"Version {Assembly.GetExecutingAssembly().GetName().Version}");
+            if (_Form.IsLoggingOn)
+                _Form.log.WriteLine("Getting Database Path");
             // Get the database path
             string databasePath = GetDatabasePath();
+
+            if (_Form.IsLoggingOn)
+                _Form.log.WriteLine($"The Database Path is {databasePath}");
 
             // Check if the DrafixUpdate.mdf file exists
             string databaseFilePath = Path.Combine(databasePath, "DrafixUpdate.mdf");
             string logFilePath = Path.Combine(databasePath, "DrafixUpdate_log.ldf");
 
+            if (_Form.IsLoggingOn)
+                _Form.log.WriteLine($"Checking to see if .mdf and .ldf files exist");
 
             if (!File.Exists(databaseFilePath))
             {
+                if (_Form.IsLoggingOn)
+                    _Form.log.WriteLine($"Could not find the mdf file");
                 _Form.UpdateStatus("Cannot find database file", 100, 100, false);
-                return false; 
+                return false;
+            }
+
+            if (!File.Exists(logFilePath))
+            {
+                if (_Form.IsLoggingOn)
+                    _Form.log.WriteLine($"Could not find the ldf file");
+                _Form.UpdateStatus("Cannot find database file", 100, 100, false);
+                return false;
             }
 
             // Check if the database is attached in SQL Server
+            if (_Form.IsLoggingOn)
+                _Form.log.WriteLine($"Checking to see if Database is Attached");
             if (!IsDatabaseAttached("DrafixUpdate"))
             {
+                if (_Form.IsLoggingOn)
+                    _Form.log.WriteLine($"The database is not attached, attempting to attach it");
                 _Form.UpdateStatus("Database 'DrafixUpdate' is not attached to SQL Server. Attempting to attach...", 100, 100, true);
 
 
@@ -46,55 +69,85 @@ namespace DatabaseUpdater
 
                 if (!attachSuccess)
                 {
+                    if (_Form.IsLoggingOn)
+                        _Form.log.WriteLine($"Uable to attach database");
                     _Form.UpdateStatus("Failed to attach the database.", 100, 100, false);
                     return false;
                 }
             }
-
+            if (_Form.IsLoggingOn)
+                _Form.log.WriteLine($"The database is attached");
             Console.WriteLine("Database is attached and ready for use.");
 
             // Execute the stored procedure
-            ExecuteStoredProcedure("DrafixUpdate", "dbo.USP_Upgrade_26_0");
-            Console.WriteLine("Uploading new objects.");
+            if (_Form.IsLoggingOn)
+                _Form.log.WriteLine($"Executing SPROC");
 
+            var didItWork = false;
+            if (ExecuteStoredProcedure("DrafixUpdate", "dbo.USP_Upgrade_26_0"))
+            {
+                didItWork = true;
+                Console.WriteLine("Uploading new objects.");
+            }
             //Close Connections
+            if (_Form.IsLoggingOn)
+                _Form.log.WriteLine($"Closing database connections");
             CloseDatabaseConnections("DrafixUpdate");
 
             //Detach DrafixUpdate Database and Deletes the DrafixUpdate MDF and LDF files
+            if (_Form.IsLoggingOn)
+                _Form.log.WriteLine($"About to detach database and delete DrafixUpdate files");
             bool detachSuccess = DetachDatabase("DrafixUpdate");
 
             if (detachSuccess)
             {
+                if (_Form.IsLoggingOn)
+                    _Form.log.WriteLine($"Database was sucessfully detached");
                 _Form.UpdateStatus("Database 'DrafixUpdate' detached successfully.", 100, 100, true);
 
-                //Delete MDF and LDF files after succesfu detachment
-                try
+                if (didItWork)
                 {
-                    if (File.Exists(databaseFilePath))
+                    //Delete MDF and LDF files after succesful detachment
+                    if (_Form.IsLoggingOn)
+                        _Form.log.WriteLine($"Deleting mdf and ldf files");
+                    try
                     {
-                        File.Delete(databaseFilePath);
-                        Console.WriteLine("Deleted DrafixUpdate.mdf");
-                    }
+                        if (File.Exists(databaseFilePath))
+                        {
+                            File.Delete(databaseFilePath);
+                            if (_Form.IsLoggingOn)
+                                _Form.log.WriteLine($"Deleted the mdf");
+                            Console.WriteLine("Deleted DrafixUpdate.mdf");
+                        }
 
-                    if (File.Exists(logFilePath))
-                    {
-                        File.Delete(logFilePath);
-                        Console.WriteLine("Deleted DrafixUpdate_log.ldf");
+                        if (File.Exists(logFilePath))
+                        {
+                            File.Delete(logFilePath);
+                            if (_Form.IsLoggingOn)
+                                _Form.log.WriteLine($"Deleted the ldf");
+                            Console.WriteLine("Deleted DrafixUpdate_log.ldf");
+                        }
                     }
-                }
-                catch (Exception ex)
-                {
-                    _Form.UpdateStatus($"Error deleting database files: {ex.Message}", 100, 100, false);
-                    return false;
+                    catch (Exception ex)
+                    {
+                        if (_Form.IsLoggingOn)
+                            _Form.log.WriteLine($"Proplem deleting database files {ex.Message}");
+                        _Form.UpdateStatus($"Error deleting database files: {ex.Message}", 100, 100, false);
+                        return false;
+                    }
                 }
             }
             else
             {
+                if (_Form.IsLoggingOn)
+                    _Form.log.WriteLine($"Failed to detach");
                 _Form.UpdateStatus("Failed to detach the database 'DrafixUpdate'.", 100, 100, false);
                 return false;
             }
 
             // Exit the application after the process is done
+            if (_Form.IsLoggingOn)
+                _Form.log.WriteLine($"Done");
             Application.Exit();
 
             return true;
@@ -184,6 +237,8 @@ namespace DatabaseUpdater
                 }
                 catch (Exception ex)
                 {
+                    if (_Form.IsLoggingOn)
+                        _Form.log.WriteLine($"Error {ex.Message}");
                     _Form.UpdateStatus("Error checking database attachment: " + ex.Message, 100, 100, false);
                 }
             }
@@ -227,6 +282,8 @@ namespace DatabaseUpdater
                 }
                 catch (Exception ex)
                 {
+                    if (_Form.IsLoggingOn)
+                        _Form.log.WriteLine($"Error {ex.Message}");
                     _Form.UpdateStatus("Error attaching the database: " + ex.Message, 100, 100, false);
 
                     return false;
@@ -239,8 +296,9 @@ namespace DatabaseUpdater
         /// </summary>
         /// <param name="databaseName">The name of the database where the stored procedure is located.</param>
         /// <param name="storedProcedureName">The name of the stored procedure to execute.</param>
-        private void ExecuteStoredProcedure(string databaseName, string storedProcedureName)
+        private bool ExecuteStoredProcedure(string databaseName, string storedProcedureName)
         {
+            bool result = false;
             //string connectionString = $@"Server=localhost\SQLEXPRESS;Database={databaseName};Trusted_Connection=True;";
             string connectionString = "DSN=DrafixSQL;Database=master;ExtendedAnsiSQL=1;TrustedConnection=True;";
 
@@ -257,14 +315,17 @@ namespace DatabaseUpdater
                         // Execute the stored procedure
                         command.ExecuteNonQuery();
                         _Form.UpdateStatus($"Stored procedure '{storedProcedureName}' executed successfully.", 100, 100, true);
-
+                        result = true;
                     }
                 }
                 catch (Exception ex)
                 {
+                    if (_Form.IsLoggingOn)
+                        _Form.log.WriteLine($"Error {ex.Message}");
                     _Form.UpdateStatus("Error executing stored procedure: " + ex.Message, 100, 100, false);
 
                 }
+                return result;
             }
         }
 
@@ -333,6 +394,8 @@ namespace DatabaseUpdater
                 }
                 catch (Exception ex)
                 {
+                    if (_Form.IsLoggingOn)
+                        _Form.log.WriteLine($"Error {ex.Message}");
                     _Form.UpdateStatus("Error detaching the database: " + ex.Message, 100, 100, false);
 
                     return false;
